requirements.txt
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
Flask==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==23.2
ply==3.11
psycopg2-binary==2.9.9
pycparser==2.21
Werkzeug==3.0.1
numpy==1.26.0